refactor(settings): split useSettingsContext out of the nav hook
Separate the gating-context primitive (useSettingsContext) from the nav
model (useSettingsNavigation, which consumes it). Rename the nav types to
SettingsNavigationItem/Section and the renderer to SettingsNavigation.
